@@ -1,5 +1,37 @@
 import 'package:flutter/material.dart';
 
-Widget addNoSpace(double height) {
-  return SizedBox(height: height);
+import 'package:flutter/material.dart';
+
+class ButtonWidget extends StatelessWidget {
+  final IconData icon;
+  final String text;
+  final VoidCallback onClicked;
+
+  const ButtonWidget({
+    Key? key,
+    required this.icon,
+    required this.text,
+    required this.onClicked,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) => ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          minimumSize: Size.fromHeight(50),
+        ),
+        child: buildContent(),
+        onPressed: onClicked,
+      );
+
+  Widget buildContent() => Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Icon(icon, size: 28),
+          SizedBox(width: 16),
+          Text(
+            text,
+            style: TextStyle(fontSize: 22, color: Colors.white),
+          ),
+        ],
+      );
 }
